Initial check-in
This is mostly a clone of https://github.com/BramRausch/PiPod with minor changes to main.py and device.py. The changes are related to hooking the GPIOs directly to the pygame que, rather than using Retrogame. Not all GPIOs hooked yet -- just the one to prove that the concept works.